+++ libafl_qemu/src/helpers/asan.rs
@@ -17,11 +17,10 @@ use num_enum::{IntoPrimitive, TryFromPrimitive};
 use rangemap::RangeMap;
 
 use crate::{
-    calls::FullBacktraceCollector,
     emu::{EmuError, MemAccessInfo, SyscallHookResult},
-    helper::{
-        HasInstrumentationFilter, IsFilter, QemuHelper, QemuHelperTuple,
-        QemuInstrumentationAddressRangeFilter,
+    helpers::{
+        calls::FullBacktraceCollector, HasInstrumentationFilter, IsFilter, QemuHelper,
+        QemuHelperTuple, QemuInstrumentationAddressRangeFilter,
     },
     hooks::{Hook, QemuHooks},
     snapshot::QemuSnapshotHelper,
