+++ libafl_qemu/src/helpers/drcov.rs
