+++ libafl_qemu/src/helpers/injections.rs
