+++ libafl_qemu/src/helpers/snapshot.rs
@@ -25,7 +25,7 @@ use crate::SYS_newfstatat;
 use crate::{
     asan::QemuAsanHelper,
     emu::SyscallHookResult,
-    helper::{QemuHelper, QemuHelperTuple},
+    helpers::{QemuHelper, QemuHelperTuple},
     hooks::{Hook, QemuHooks},
     Qemu, SYS_fstat, SYS_fstatfs, SYS_futex, SYS_getrandom, SYS_mprotect, SYS_mremap, SYS_munmap,
     SYS_pread64, SYS_read, SYS_readlinkat, SYS_statfs,
