+++ libafl_qemu/src/lib.rs
@@ -33,80 +33,16 @@ use std::env;
 pub use libafl_qemu_sys as sys;
 pub use strum::IntoEnumIterator;
 
-#[cfg(cpu_target = "aarch64")]
-pub mod aarch64;
-#[cfg(all(cpu_target = "aarch64", not(feature = "clippy")))]
-pub use aarch64::*;
-
-#[cfg(cpu_target = "arm")]
-pub mod arm;
-#[cfg(all(cpu_target = "arm", not(feature = "clippy")))]
-pub use arm::*;
-
-#[cfg(cpu_target = "i386")]
-pub mod i386;
-#[cfg(all(cpu_target = "i386", not(feature = "clippy")))]
-pub use i386::*;
-
-#[cfg(cpu_target = "x86_64")]
-pub mod x86_64;
-#[cfg(cpu_target = "x86_64")]
-pub use x86_64::*;
-
-#[cfg(cpu_target = "mips")]
-pub mod mips;
-#[cfg(cpu_target = "mips")]
-pub use mips::*;
-
-#[cfg(cpu_target = "ppc")]
-pub mod ppc;
-#[cfg(cpu_target = "ppc")]
-pub use ppc::*;
-
-#[cfg(cpu_target = "hexagon")]
-pub mod hexagon;
-#[cfg(cpu_target = "hexagon")]
-pub use hexagon::*;
+pub mod arch;
+pub use arch::*;
 
 pub mod elf;
 
-pub mod helper;
-pub use helper::*;
+pub mod helpers;
+pub use helpers::*;
+
 pub mod hooks;
 pub use hooks::*;
-
-pub mod edges;
-pub use edges::QemuEdgeCoverageHelper;
-
-#[cfg(not(any(cpu_target = "mips", cpu_target = "hexagon")))]
-pub mod cmplog;
-#[cfg(not(any(cpu_target = "mips", cpu_target = "hexagon")))]
-pub use cmplog::QemuCmpLogHelper;
-
-#[cfg(all(emulation_mode = "usermode", feature = "injections"))]
-pub mod injections;
-#[cfg(all(emulation_mode = "usermode", feature = "injections"))]
-pub use injections::QemuInjectionHelper;
-
-#[cfg(all(emulation_mode = "usermode", not(cpu_target = "hexagon")))]
-pub mod snapshot;
-#[cfg(all(emulation_mode = "usermode", not(cpu_target = "hexagon")))]
-pub use snapshot::QemuSnapshotHelper;
-
-#[cfg(all(emulation_mode = "usermode", not(cpu_target = "hexagon")))]
-pub mod asan;
-#[cfg(all(emulation_mode = "usermode", not(cpu_target = "hexagon")))]
-pub use asan::{init_qemu_with_asan, QemuAsanHelper};
-
-#[cfg(all(emulation_mode = "usermode", not(cpu_target = "hexagon")))]
-pub mod asan_guest;
-#[cfg(all(emulation_mode = "usermode", not(cpu_target = "hexagon")))]
-pub use asan_guest::{init_qemu_with_asan_guest, QemuAsanGuestHelper};
-
-#[cfg(not(cpu_target = "hexagon"))]
-pub mod calls;
-#[cfg(not(cpu_target = "hexagon"))]
-pub mod drcov;
 
 pub mod executor;
 pub use executor::QemuExecutor;
@@ -118,7 +54,7 @@ pub use emu::*;
 
 pub mod breakpoint;
 pub mod command;
-pub mod sync_backdoor;
+pub mod sync_exit;
 
 #[must_use]
 pub fn filter_qemu_args() -> Vec<String> {
