+++ libafl_qemu/src/sync_exit.rs
